Simplify the Matcher trait
Now matchers need to implement a single `description` method, which will
return all the necessary information to build any kind of failure
message.

This removes the duplication between the old `failure_message` and
`negated_failure_message` and leaves the formatting logic for failure
messages in one single place, the `Expectation`.

I anticipate `Description` becoming more and more complex to accomodate
all sorts of failures, but we should be able to make it flexible enough
to accomodate all matchers.